requirements.txt
beautifulsoup4>=4.12.0
requests>=2.28.2
langchain>=0.1.0
langchain-google-genai>=0.0.5
langchain-core>=0.1.0
langchain-community>=0.0.10
chromadb>=0.4.18
google-generativeai>=0.3.1
rich>=13.5.0 